* test/openssl/test_ssl.rb: add workaround for Cygwin.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_8@6674 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ test/openssl/test_ssl.rb
@@ -70,7 +70,7 @@ class OpenSSL::TestSSL < Test::Unit::TestCase
       block.call(server)
     ensure
       if server
-        Process.kill(:KILL, pid)
+        Process.kill(:KILL, pid) if pid > 0
         server.close
       end
     end
@@ -78,6 +78,10 @@ class OpenSSL::TestSSL < Test::Unit::TestCase
 
   def starttls(ssl)
     ssl.puts("STARTTLS")
+
+    sleep 1   # When this line is eliminated, process on Cygwin blocks
+              # forever at ssl.connect. But I don't know why it does.
+
     ssl.connect
   end
